requirements.txt
python-dotenv==0.21.0
telegram
python-telegram-bot==13.7
python-dotenv==0.21.0